Seperate completed functions to a different js file
